+++ sklearn/grid_search.py
@@ -329,10 +329,10 @@ class GridSearchCV(BaseEstimator, MetaEstimatorMixin):
         self.pre_dispatch = pre_dispatch
 
     def _set_methods(self):
-        if hasattr(self._best_estimator_, 'predict'):
-            self.predict = self._best_estimator_.predict
-        if hasattr(self._best_estimator_, 'predict_proba'):
-            self.predict_proba = self._best_estimator_.predict_proba
+        if hasattr(self.best_estimator_, 'predict'):
+            self.predict = self.best_estimator_.predict
+        if hasattr(self.best_estimator_, 'predict_proba'):
+            self.predict_proba = self.best_estimator_.predict_proba
 
     def fit(self, X, y=None, **params):
         """Run fit with all sets of parameters
@@ -379,7 +379,7 @@ class GridSearchCV(BaseEstimator, MetaEstimatorMixin):
             params = next(iter(grid))
             base_clf.set_params(**params)
             base_clf.fit(X, y)
-            self._best_estimator_ = base_clf
+            self.best_estimator_ = base_clf
             self._set_methods()
             return self
 
@@ -434,7 +434,7 @@ class GridSearchCV(BaseEstimator, MetaEstimatorMixin):
             # clone first to work around broken estimators
             best_estimator = clone(base_clf).set_params(**best_params)
             best_estimator.fit(X, y, **self.fit_params)
-            self._best_estimator_ = best_estimator
+            self.best_estimator_ = best_estimator
             self._set_methods()
 
         # Store the computed scores
@@ -455,14 +455,3 @@ class GridSearchCV(BaseEstimator, MetaEstimatorMixin):
                              % self.best_estimator_)
         y_predicted = self.predict(X)
         return self.score_func(y, y_predicted)
-
-    # TODO around 0.13: remove this property, make it an attribute
-    @property
-    def best_estimator_(self):
-        if hasattr(self, '_best_estimator_'):
-            return self._best_estimator_
-        else:
-            raise RuntimeError("Grid search has to be run with 'refit=True'"
-                " to make predictions or obtain an instance  of the best "
-                " estimator. To obtain the best parameter settings, "
-                " use ``best_params_``.")