+++ sklearn/linear_model/__init__.py
@@ -26,7 +26,6 @@ from .omp import orthogonal_mp, orthogonal_mp_gram, OrthogonalMatchingPursuit
 from .perceptron import Perceptron
 from .randomized_l1 import RandomizedLasso, RandomizedLogisticRegression, \
                            lasso_stability_path
-from . import sparse
 from .isotonic_regression_ import IsotonicRegression
 
 __all__ = ['ARDRegression',
@@ -65,5 +64,4 @@ __all__ = ['ARDRegression',
            'lasso_stability_path',
            'orthogonal_mp',
            'orthogonal_mp_gram',
-           'ridge_regression',
-           'sparse']
+           'ridge_regression']