+++ sklearn/mixture/gmm.py
@@ -308,31 +308,6 @@ class GMM(BaseEstimator):
         responsibilities = np.exp(lpr - logprob[:, np.newaxis])
         return logprob, responsibilities
 
-    @deprecated("""will be removed in v0.13;
-    use the score or predict method instead, depending on the question""")
-    def decode(self, X):
-        """Find most likely mixture components for each point in X.
-
-        DEPRECATED IN VERSION 0.11; WILL BE REMOVED IN VERSION 0.13.
-        use the score or predict method instead, depending on the question.
-
-        Parameters
-        ----------
-        X : array_like, shape (n, n_features)
-            List of n_features-dimensional data points.  Each row
-            corresponds to a single data point.
-
-        Returns
-        -------
-        logprobs : array_like, shape (n_samples,)
-            Log probability of each point in `obs` under the model.
-
-        components : array_like, shape (n_samples,)
-            Index of the most likelihod mixture components for each observation
-        """
-        logprob, posteriors = self.eval(X)
-        return logprob, posteriors.argmax(axis=1)
-
     def score(self, X):
         """Compute the log probability under the model.
 
@@ -381,16 +356,6 @@ class GMM(BaseEstimator):
         logprob, responsibilities = self.eval(X)
         return responsibilities
 
-    @deprecated("""will be removed in v0.13;
-    use the score or predict method instead, depending on the question""")
-    def rvs(self, n_samples=1, random_state=None):
-        """Generate random samples from the model.
-
-        DEPRECATED IN VERSION 0.11; WILL BE REMOVED IN VERSION 0.12
-        use sample instead
-        """
-        return self.sample(n_samples, random_state)
-
     def sample(self, n_samples=1, random_state=None):
         """Generate random samples from the model.
 
@@ -431,7 +396,7 @@ class GMM(BaseEstimator):
                     num_comp_in_X, random_state=random_state).T
         return X
 
-    def fit(self, X, **kwargs):
+    def fit(self, X):
         """Estimate model parameters with the expectation-maximization
         algorithm.
 
@@ -455,24 +420,6 @@ class GMM(BaseEstimator):
             raise ValueError(
                 'GMM estimation with %s components, but got only %s samples' %
                 (self.n_components, X.shape[0]))
-        if kwargs:
-            warnings.warn("Setting parameters in the 'fit' method is"
-                          "deprecated and will be removed in 0.13. Set it on "
-                          "initialization instead.", DeprecationWarning,
-                          stacklevel=2)
-            # initialisations for in case the user still adds parameters to fit
-            # so things don't break
-            if 'n_iter' in kwargs:
-                self.n_iter = kwargs['n_iter']
-            if 'n_init' in kwargs:
-                if kwargs['n_init'] < 1:
-                    raise ValueError('GMM estimation requires n_init > 0.')
-                else:
-                    self.n_init = kwargs['n_init']
-            if 'params' in kwargs:
-                self.params = kwargs['params']
-            if 'init_params' in kwargs:
-                self.init_params = kwargs['init_params']
 
         max_log_prob = -np.infty